removing unncessary float

+++ hummingbird/ml/operator_converters/_binarizer_implementations.py
@@ -24,4 +24,4 @@ class Binarizer(BaseOperator, torch.nn.Module):
         self.threshold = torch.nn.Parameter(torch.FloatTensor([threshold]), requires_grad=False)
 
     def forward(self, x):
-        return torch.gt(x, self.threshold).float()
+        return torch.gt(x, self.threshold)
